Remove some redundant and annotate timer functions
