+++ src/main.rs
@@ -1,5 +1,5 @@
-#![allow(dead_code)]
-#![allow(unused)]
+// #![allow(dead_code)]
+// #![allow(unused)]
 
 mod cli;
 mod debugger;
